updated scraper to have efficient duplicate lookup and changed lat/long storing type

+++ packages/db/migrations/0016_map_locations.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS "map_location" (
 	"id" varchar PRIMARY KEY NOT NULL,
 	"name" varchar NOT NULL,
-	"latitude" numeric NOT NULL,
-	"longitude" numeric NOT NULL
+	"latitude" double precision NOT NULL,
+	"longitude" double precision NOT NULL
 );
